gst_v4l2_buffer_pool_poll: Continue if v4l FD has error
There are two error cases here:

1. The `poll` fails, setting `errno`.  In this case we get a useful error
   message including `strerror(errno)` and streaming stops
2. The `poll` succeeds but the v4l FD has an error set on it.  In this
   case `errno` is not set, but `strerror(errno)` was still reported.
   This is misleading because `errno` is unrelated to the reason the
   message was printed.

Now in the case of (2) we warn, but continue as if polling had succeeded.
The theory is that the code will attempt an operation on the v4l FD and
a real error will be reported then which will relevant and enable
debugging the real cause of the failure.